backend/requirements.txt
packaging==23.1  # Necessário para o flash_attn
accelerate==0.21.0  # Adiciona suporte ao uso de device_map e otimização de memória
#flash_attn==2.5.8
numpy==1.24.4
Pillow==10.3.0
requests==2.31.0
torch==2.3.0
torchvision==0.18.0
transformers==4.40.2
Flask
flask-cors
